+++ NEWS.md
@@ -128,6 +128,7 @@ New library features
 certain compiler plugin workflows ([#56660]).
 * `sort` now supports `NTuple`s ([#54494])
 * `map!(f, A)` now stores the results in `A`, like `map!(f, A, A)`. or `A .= f.(A)` ([#40632]).
+* `Timer` now has readable `timeout` and `interval` properties, and a more descriptive show method ([#57081])
 
 Standard library changes
 ------------------------